refactor: Mapper 경로 구조 정리
