+++ src/main/java/com/antmillion/communication/controller/CommunityController.java
@@ -3,6 +3,7 @@ package com.antmillion.communication.controller;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -12,9 +13,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+
 import com.antmillion.communication.dto.CommunityDTO;
 import com.antmillion.communication.dto.CommunityDetailResponseDTO;
 import com.antmillion.communication.service.CommunityService;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
