+++ src/main/java/com/antmillion/communication/mapper/CommunityMapper.java
@@ -1,4 +1,4 @@
-package com.antmillion.mappers;
+package com.antmillion.communication.mapper;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
