--- src/main/java/com/antmillion/mappers/TestMapper.java
@@ -1,5 +1,0 @@
-package com.antmillion.mappers;
-
-public interface TestMapper {
-
-}
