+++ src/main/resources/mappers/communication/CommunityMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
     "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.antmillion.mappers.CommunityMapper">
+<mapper namespace="com.antmillion.communication.mapper.CommunityMapper">
 
     <!-- ResultMap 정의 -->
     <resultMap id="CommunityDetailResponseMap" type="com.antmillion.communication.dto.CommunityDetailResponseDTO">
@@ -15,7 +15,7 @@
         <result property="postedDate" column="posted_date"/>
     </resultMap>
 
-    <!-- 특정 종목의 커뮤니티 글 목록 조회-->
+    <!-- 특정 종목의 커뮤니티 글 목록 조회 (최신순) -->
     <select id="selectCommunityListByStock" resultMap="CommunityDetailResponseMap">
         SELECT 
             c.post_id,
